Revamped Class layout - everything now uses main.Engine
