+++ bamboo/cooling.py
@@ -17,7 +17,7 @@ import bamboo as bam
 import numpy as np
 import matplotlib.pyplot as plt
 import scipy
-import json
+
 
 #Check if CoolProp is installed
 import imp
@@ -135,120 +135,6 @@ def h_coolant_1(A, D, mdot, mu, k, c_bar, rho):
     return 0.023*c_bar * (mdot/A) * (D*v*rho/mu)**(-0.2) * (mu*c_bar/k)**(-2/3)
 
 
-class EngineGeometry:
-    """Class for storing and calculating features of the engine's geometry.
-
-    Args:
-        nozzle (float): Nozzle of the engine.
-        chamber_length (float): Length of the combustion chamber (m)
-        chamber_area (float): Cross sectional area of the combustion chamber (m^2)
-        wall_thickness (float or array): Thickness of the inner liner wall (m). Can be constant (float), or variable (array).
-        geometry (str, optional): Geometry system to use. Currently the only option is 'auto'. Defaults to "auto".
-
-    """
-    def __init__(self, nozzle, chamber_length, chamber_area, wall_thickness, geometry="auto"):
-        self.nozzle = nozzle
-        self.chamber_length = chamber_length
-        self.chamber_area = chamber_area
-        self.chamber_radius = (chamber_area/np.pi)**0.5 
-        if type(wall_thickness) is (float or int):
-            self.wall_thickness = [wall_thickness]  #Convert into a list so the interpolation works
-        else:
-            self.wall_thickness = wall_thickness
-        self.geometry = geometry
-
-        if self.nozzle.At > self.chamber_area:
-            raise ValueError(f"The combustion chamber area {self.chamber_area} m^2 is smaller than the throat area {self.nozzle.At} m^2.")
-
-        if self.geometry == "auto":
-            #Use the system defined in Reference [1] - mostly using Eqns (4)
-            #Make sure we cap the size of the converging section to the radius of the combustion chamber.
-            chamber_radius = (self.chamber_area/np.pi)**0.5
-            theta_min = -np.pi - np.arcsin((chamber_radius - self.nozzle.Rt - 1.5*self.nozzle.Rt) / (1.5*self.nozzle.Rt)) 
-            if theta_min > -3*np.pi/4:
-                self.theta_curved_converging_start = theta_min
-            else:
-                self.theta_curved_converging_start = -3*np.pi/4
-
-            #Find key properties for the converging section
-            self.x_curved_converging_start = 1.5*self.nozzle.Rt*np.cos(self.theta_curved_converging_start)
-            self.y_curved_converging_start = 1.5*self.nozzle.Rt*np.sin(self.theta_curved_converging_start) + 1.5*self.nozzle.Rt + self.nozzle.Rt
-
-            #Find the gradient where the curved converging bit starts
-            dxdtheta_curved_converging_start = -1.5*self.nozzle.Rt*np.sin(self.theta_curved_converging_start)
-            self.dydx_curved_converging_start = -1.5*self.nozzle.Rt*np.cos(self.theta_curved_converging_start)/dxdtheta_curved_converging_start
-
-            #Find the x-position where we reach the combustion chamber radius
-            self.x_chamber_end = self.x_curved_converging_start - (self.chamber_radius - self.y_curved_converging_start)/self.dydx_curved_converging_start
-
-            #Start and end points of the engine
-            self.x_min = self.x_chamber_end - self.chamber_length
-            self.x_max = self.nozzle.length
-
-    def y(self, x):
-        """Get the radius of the engine contour for a given x position.
-
-        Args:
-            x (float): x position (m). x = 0 is the throat, x > 0 is the nozzle diverging section.
-
-        Returns:
-            float: Radius of the engine contour (m).
-        """
-        if self.geometry == "auto":
-            #Curved converging section
-            if x < 0 and x > self.x_curved_converging_start:
-                theta = -np.arccos(x/(1.5*self.nozzle.Rt))
-                return 1.5*self.nozzle.Rt*np.sin(theta) + 1.5*self.nozzle.Rt + self.nozzle.Rt
-
-            #Before the curved part of the converging section
-            elif x <= self.x_curved_converging_start:
-                #Inside the chamber
-                if x < self.x_chamber_end and x >= self.x_min:
-                    return self.chamber_radius
-
-                #Inside the converging section
-                elif x >= self.x_chamber_end:
-                    return np.interp(x, [self.x_chamber_end, self.x_curved_converging_start], [self.chamber_radius, self.y_curved_converging_start])
-
-                #Outside of the engine
-                else:
-                    return ValueError(f"x is beyond the front of the engine. You tried to input {x} but the minimum value you're allowed is {self.x_chamber_end - self.chamber_length}")
-            
-            #In the diverging section of the nozzle
-            elif x >= 0:
-                return self.nozzle.y(x)
-
-    def A(self, x):
-        """Get the engine cross sectional area at a given x position.
-
-        Args:
-            x (float): x position (m). x = 0 is the throat, x > 0 is the nozzle diverging section.
-
-        Returns:
-            float: Cross sectional area (m^2)
-        """
-        return np.pi*self.y(x)**2
-
-    def plot_geometry(self, number_of_points = 1000):
-        """Plots the engine geometry. Note that to see the plot, you will need to run matplotlib.pyplot.show().
-
-        Args:
-            number_of_points (int, optional): Numbers of discrete points to plot. Defaults to 1000.
-        """
-        x = np.linspace(self.x_min, self.x_max, number_of_points)
-        y = np.zeros(len(x))
-
-        for i in range(len(x)):
-            y[i] = self.y(x[i])
-
-        fig, axs = plt.subplots()
-        axs.plot(x, y, color="blue")
-        axs.plot(x, -y, color="blue")
-        axs.grid()
-        axs.set_aspect('equal')
-        plt.xlabel("x position (m)")
-        plt.ylabel("y position (m)")
-
 class CoolingJacket:
     """Cooling jacket parameters.
 
@@ -258,6 +144,7 @@ class CoolingJacket:
         inlet_p0 (float): Inlet coolant stagnation pressure (Pa)
         coolant_transport (TransportProperties): Container for the coolant transport properties.
         mdot_coolant (float): Coolant mass flow rate (kg/s)
+        xs (list): x position that the cooling jacket starts and ends at in the form [x_start, x_end]. Defaults to [-1000, 1000].
         channel_shape (str, optional): Options include 'rectangle', 'semi-circle', and 'custom'. Defaults to "rectangle".
         configuration (str, optional): Options include 'spiral'. Defaults to "spiral".
     
@@ -268,11 +155,12 @@ class CoolingJacket:
         custom_effective_diameter (float, optional) : If using channel_shape = 'custom', this is the effective diameter you want to use. 
         custom_flow_area (float, optional) : If using channel_shape = 'custom', this is the flow you want to use. 
     """
-    def __init__(self, inner_wall, inlet_T, inlet_p0, coolant_transport, mdot_coolant, channel_shape = "rectangle", configuration = "spiral", **kwargs):
+    def __init__(self, inner_wall, inlet_T, inlet_p0, coolant_transport, mdot_coolant, xs = [-1000, 1000], channel_shape = "rectangle", configuration = "spiral", **kwargs):
 
         self.inner_wall = inner_wall
         self.coolant_transport = coolant_transport       
         self.mdot_coolant = mdot_coolant
+        self.xs = xs
         self.inlet_T = inlet_T
         self.inlet_p0 = inlet_p0
         self.channel_shape = channel_shape
@@ -330,6 +218,18 @@ class CoolingJacket:
             float: Coolant stagnation pressure
         """
         return self.inlet_p0
+
+    def coolant_velocity(self, x, rho_coolant):
+        """Get coolant velocity using mdot = rho*V*A.
+
+        Args:
+            x (float): Axial position
+            rho_coolant (float): Coolant density (kg/m3)
+
+        Returns:
+            float: Coolant velocity (m/s)
+        """
+        return self.mdot_coolant/(rho_coolant * self.A(x))
 
 class Material:
     """Class used to specify a material and its properties. 
@@ -489,371 +389,6 @@ class TransportProperties:
         elif self.model == "CoolProp":
             return PropsSI("DMASS", "T", T, "P", p, self.coolprop_name)
 
-class EngineWithCooling:
-    """Used for running cooling system analyses.
-
-    Args:
-        chamber_conditions (ChamberConditions): Engine chamber conditions object.
-        geometry (EngineGeometry): Engine geometry.
-        cooling_jacket (CoolingJacket): Cooling jacket properties.
-        perfect_gas (PerfectGas): Properties of the exhaust gas.
-        exhaust_transport (TransportProperties): Container for the exhaust gas transport properties.
-    """
-    def __init__(self, chamber_conditions, geometry, cooling_jacket, perfect_gas, exhaust_transport):
-        self.chamber_conditions = chamber_conditions
-        self.geometry = geometry
-        self.cooling_jacket = cooling_jacket
-        self.perfect_gas = perfect_gas
-        self.exhaust_transport = exhaust_transport
-
-        self.c_star = self.chamber_conditions.p0 * self.geometry.nozzle.At / self.chamber_conditions.mdot
-
-        #Check if the nozzle is choked
-        max_throat_area = bam.get_throat_area(perfect_gas, chamber_conditions)
-        if self.geometry.nozzle.At > max_throat_area:
-            raise ValueError(f"The nozzle throat is not choked. You need to reduce the throat area to at least {max_throat_area} m^2")
-
-    def M(self, x):
-        """Get exhaust gas Mach number.
-
-        Args:
-            x (float): Axial position along the engine (m). Throat is at x = 0.
-
-        Returns:
-            float: Mach number of the freestream.
-        """
-        #If we're at the throat M=1 by default:
-        if x==0:
-            return 1.00
-
-        #If we're not at the throat:
-        else:
-            def func_to_solve(Mach):
-                return self.chamber_conditions.mdot*(self.perfect_gas.cp*self.chamber_conditions.T0)**0.5 / (self.geometry.A(x)*self.chamber_conditions.p0) - bam.m_bar(Mach, self.perfect_gas.gamma)
-            
-            if x > 0:
-                Mach = scipy.optimize.root_scalar(func_to_solve, bracket = [1,300], x0 = 1).root
-            else:
-                Mach = scipy.optimize.root_scalar(func_to_solve, bracket = [0,1], x0 = 0.5).root
-
-            return Mach
-
-    def T(self, x):
-        """Get exhaust gas temperature.
-
-        Args:
-            x (float): Axial position (m). Throat is at x = 0.
-
-        Returns:
-            float: Temperature (K)
-        """
-        return bam.T(self.chamber_conditions.T0, self.M(x), self.perfect_gas.gamma)
-
-    def p(self, x):
-        """Get exhaust gas pressure.
-
-        Args:
-            x (float): Axial position (m). Throat is at x = 0.
-
-        Returns:
-            float: Freestream pressure (Pa)
-        """
-        return bam.p(self.chamber_conditions.p0, self.M(x), self.perfect_gas.gamma)
-
-    def rho(self, x):
-        """Get exhaust gas density.
-
-        Args:
-            x (float): Axial position. Throat is at x = 0.
-
-        Returns:
-            float: Freestream gas density (kg/m^3)
-        """
-        #p = rhoRT for an ideal gas, so rho = p/RT
-        return self.p(x)/(self.T(x)*self.perfect_gas.R)
-
-    def show_gas_temperature(self, number_of_points=1000):
-        """Plot freestream gas temperature against position. Note that to see the plot, you will need to run matplotlib.pyplot.show().
-
-        Args:
-            number_of_points (int, optional): Number of points to discretise the plot into. Defaults to 1000.
-        """
-        x = np.linspace(self.geometry.x_min, self.geometry.x_max, number_of_points)
-        y = np.zeros(len(x))
-        T = np.zeros(len(x))
-
-        for i in range(len(x)):
-            y[i] = self.geometry.y(x[i])
-            T[i] = self.T(x[i])
-
-        fig, ax_shape = plt.subplots()
-
-        ax_shape.plot(x, y, color="blue")
-        ax_shape.plot(x, -y, color="blue")
-        ax_shape.set_aspect('equal')
-        ax_shape.set_xlabel("x position (m)")
-        ax_shape.set_ylabel("y position (m)")
-
-        ax_temp = ax_shape.twinx()
-        ax_temp.plot(x, T, color="orange")
-        ax_temp.grid()
-        ax_temp.set_ylabel("Temperature (K)")
-
-    def show_gas_mach(self, number_of_points=1000):
-        """Plot Mach number against position. Note that to see the plot, you will need to run matplotlib.pyplot.show().
-
-        Args:
-            number_of_points (int, optional): Number of points to discretise the plot into. Defaults to 1000.
-        """
-        x = np.linspace(self.geometry.x_min, self.geometry.x_max, number_of_points)
-        y = np.zeros(len(x))
-        M = np.zeros(len(x))
-
-        for i in range(len(x)):
-            y[i] = self.geometry.y(x[i])
-            M[i] = self.M(x[i])
-
-        fig, ax_shape = plt.subplots()
-
-        ax_shape.plot(x, y, color="blue")
-        ax_shape.plot(x, -y, color="blue")
-        ax_shape.set_aspect('equal')
-        ax_shape.set_xlabel("x position (m)")
-        ax_shape.set_ylabel("y position (m)")
-
-        ax_temp = ax_shape.twinx()
-        ax_temp.plot(x, M, color="green")
-        ax_temp.grid()
-        ax_temp.set_ylabel("Mach number")
-
-    def coolant_velocity(self, x, rho_coolant):
-        """Get coolant velocity
-
-        Args:
-            x (float): Axial position
-            rho_coolant (float): Coolant density (kg/m3)
-
-        Returns:
-            float: Coolant velocity (m/s)
-        """
-        return self.cooling_jacket.mdot_coolant/(rho_coolant * self.cooling_jacket.A(x))
-
-    def map_liner_profile(self, number_of_points=1000):
-        """Maps the provided liner thickness profile to the engine geometry,
-           so each element in cooling analysis has a thickness value.
-           
-           Args:
-                number_of_points (int): Number of discrete liner positions
-
-           Returns:
-                liner (array): Interpolated liner thickness profile
-           """
-        liner = np.zeros(number_of_points)
-        for i in range(number_of_points):
-            x_pos = i*self.geometry.chamber_length/number_of_points
-            # How far along the engine is the current point
-            liner_index = x_pos * len(self.geometry.wall_thickness)
-            liner[i] = np.interp(liner_index, range(len(self.geometry.wall_thickness)), self.geometry.wall_thickness)
-        return liner
-
-    def thermal_circuit(self, x, h_gas, h_coolant, inner_wall, wall_thickness, T_gas, T_coolant):
-        """
-        q is per unit length along the nozzle wall (axially) - positive when heat is flowing to the coolant.  
-        q_Adot is the heat flux per unit area along the nozzle wall.  
-        Uses the idea of thermal circuits and resistances - we have three resistors in series.
-
-        Args:
-            x (float): x position (m)
-            h_gas (float): Gas side convective heat transfer coefficient
-            h_coolant (float): Coolant side convective heat transfer coefficient
-            inner_wall (Material): Material object for the inner wall, needed for thermal conductivity
-            wall_thickness (float): Thickness of the inner wall at x position (m)
-            T_gas (float): Free stream gas temperature (K)
-            T_coolant (float): Coolant temperature (K)
-
-        Returns:
-            float, float, float, float, float: q_dot, R_gas, R_wall, R_coolant, q_Adot
-        """
-
-        r = self.geometry.y(x)
-        
-        r_out = r + wall_thickness
-        r_in = r 
-
-        A_in = 2*np.pi*r_out    #Inner area per unit length (i.e. just the inner circumference)
-        A_out = 2*np.pi*r_in    #Outer area per unit length (i.e. just the outer circumference)
-
-        R_gas = 1/(h_gas*A_in)
-        R_wall = np.log(r_out/r_in)/(2*np.pi*inner_wall.k)
-        R_coolant = 1/(h_coolant*A_out)
-
-        q_dot = (T_gas - T_coolant)/(R_gas + R_wall + R_coolant)    #Heat flux per unit length
-        q_Adot = q_dot / A_in                                       #Heat flux per unit area
-
-        return q_dot, R_gas, R_wall, R_coolant, q_Adot
-
-    def run_heating_analysis(self, number_of_points=1000, h_gas_model = "1", h_coolant_model = "1", to_json = "heating_output.json"):
-        """Run a simulation of the engine cooling system to get wall temperatures, coolant temperatures, etc.
-
-        Args:
-            number_of_points (int, optional): Number of discrete points to divide the engine into. Defaults to 1000.
-            h_gas_model (str, optional): Equation to use for the gas side convective heat transfer coefficients. Options are '1', '2' and '3'. Defaults to "1".
-            h_coolant_model (str, optional): Equation to use for the coolant side convective heat transfer coefficients, currently the only option is '1'. Defaults to "1".
-            to_json (str or bool, optional): Directory to export a .JSON file to, containing simulation results. If False, no .JSON file is saved. Defaults to 'heating_output.json'.
-
-        Returns:
-            dict: Results of the simulation. Contains the following dictionary keys: 
-                - "x" : x positions corresponding to the rest of the data (m)
-                - "T_wall_inner" : Exhaust gas side wall temperature (K)
-                - "T_wall_outer" : Coolant side wall temperature (K)
-                - "T_coolant" : Coolant temperature (K)
-                - "T_gas" : Exhaust gas freestream temperature (K)
-                - "q_dot" : Heat transfer rate per unit length (axially along the engine) (W/m)
-                - "q_Adot": Heat transfer rate per unit area (W/m^2)
-                - "h_gas" : Convective heat transfer rate for the exhaust gas side
-                - "h_coolant" : Convective heat transfer rate for the coolant side
-                - "boil_off_position" : x position of any coolant boil off. Equal to None if the coolant does not boil.
-        """
-
-        '''Initialise variables and arrays'''
-        #To keep track of any coolant boiling
-        boil_off_position = None
-        
-        #Discretisation of the nozzle
-        discretised_x = np.linspace(self.geometry.x_max, self.geometry.x_min, number_of_points) #Run from the back end (the nozzle exit) to the front (chamber)
-        dx = discretised_x[0] - discretised_x[1]
-
-        #Discretised liner thickness
-        liner = self.map_liner_profile(number_of_points)
-
-        #Temperatures and heat transfer rates
-        T_wall_inner = np.zeros(len(discretised_x)) #Gas side wall temperature
-        T_wall_outer = np.zeros(len(discretised_x)) #Coolant side wall temperature
-        T_coolant = np.zeros(len(discretised_x))    #Coolant temperature
-        T_gas = np.zeros(len(discretised_x))        #Freestream gas temperature
-        q_dot = np.zeros(len(discretised_x))        #Heat transfer rate per unit length
-        q_Adot = np.zeros(len(discretised_x))       #Heat transfer rate per unit area
-        h_gas = np.zeros(len(discretised_x))
-        h_coolant = np.zeros(len(discretised_x))
-
-        '''Main loop'''
-        for i in range(len(discretised_x)):
-            x = discretised_x[i]
-            T_gas[i] = self.T(x)
-
-            #Calculate the current coolant temperature
-            if i == 0:
-                T_coolant[i] = self.cooling_jacket.inlet_T
-
-            else:
-                #Increase in coolant temperature, q*dx = mdot*Cp*dT
-                T_coolant[i] = T_coolant[i-1] + (q_dot[i-1]*dx)/(self.cooling_jacket.mdot_coolant*cp_coolant)   
-
-            #Update coolant heat capacity
-            cp_coolant = self.cooling_jacket.coolant_transport.cp(T = T_coolant[i], p = self.cooling_jacket.p0(x))
-
-            #Gas side heat transfer coefficient
-            if h_gas_model == "1":
-                h_gas[i] = h_gas_1(2*self.geometry.y(x),
-                                   self.M(x),
-                                   T_gas[i],
-                                   self.rho(x),
-                                   self.perfect_gas.gamma,
-                                   self.perfect_gas.R,
-                                   self.exhaust_transport.mu(T = T_gas[i], p = self.p(x)),
-                                   self.exhaust_transport.k(T = T_gas[i], p = self.p(x)),
-                                   self.exhaust_transport.Pr(T = T_gas[i], p = self.p(x)))
-
-            elif h_gas_model == "2":
-                gamma = self.perfect_gas.gamma
-                R = self.perfect_gas.R
-                D = 2*self.geometry.y(x)            #Flow diameter
-
-                #Freestream properties
-                p_inf = self.p(x)
-                T_inf = T_gas[i]
-                rho_inf = self.rho(x)
-                M_inf = self.M(x)
-                v_inf = M_inf * (gamma*R*T_inf)**0.5    #Gas velocity
-                mu_inf = self.exhaust_transport.mu(T = T_gas[i], p = p_inf)
-                Pr_inf = self.exhaust_transport.Pr(T = T_gas[i], p = p_inf)
-                cp_inf = self.perfect_gas.cp
-
-                #Properties at arithmetic mean of T_wall and T_inf
-                T_am = (T_inf + T_wall_inner[i-1]) / 2
-                mu_am = self.exhaust_transport.mu(T = T_am, p = p_inf)
-                rho_am = p_inf/(R*T_am)                                 #p = rho R T - pressure is roughly uniform across the boundary layer so p_inf ~= p_wall
-
-                #Stagnation properties
-                p0 = self.chamber_conditions.p0
-                T0 = self.chamber_conditions.T0
-                mu0 = self.exhaust_transport.mu(T =  T0, p = p0)
-
-                h_gas[i] = h_gas_2(D, cp_inf, mu_inf, Pr_inf, rho_inf, v_inf, rho_am, mu_am, mu0)
-
-            elif h_gas_model == "3":
-                h_gas[i] = h_gas_3(self.c_star,
-                                   self.geometry.nozzle.At, 
-                                   self.geometry.A(x), 
-                                   self.chamber_conditions.p0, 
-                                   self.chamber_conditions.T0, 
-                                   self.M(x), 
-                                   T_wall_inner[i-1], 
-                                   self.exhaust_transport.mu(T = T_gas[i], p = self.p(x)), 
-                                   self.perfect_gas.cp, 
-                                   self.perfect_gas.gamma, 
-                                   self.exhaust_transport.Pr(T = T_gas[i], p = self.p(x)))
-
-            else:
-                raise AttributeError(f"Could not find the h_gas_model '{h_gas_model}'")
-            
-            #Coolant side heat transfer coefficient
-            if h_coolant_model == "1":
-                h_coolant[i] = h_coolant_1(self.cooling_jacket.A(x), 
-                                           self.cooling_jacket.D(x), 
-                                           self.cooling_jacket.mdot_coolant, 
-                                           self.cooling_jacket.coolant_transport.mu(T = T_coolant[i], p = self.cooling_jacket.p0(x)), 
-                                           self.cooling_jacket.coolant_transport.k(T = T_coolant[i], p = self.cooling_jacket.p0(x)), 
-                                           cp_coolant, 
-                                           self.cooling_jacket.coolant_transport.rho(T = T_coolant[i], p = self.cooling_jacket.p0(x)))
-
-            else:
-                raise AttributeError(f"Could not find the h_coolant_model '{h_coolant_model}'")
-            
-            #Check for coolant boil off - a CoolProp uses a phase index of '0' to refer to the liquid state (see http://www.coolprop.org/coolprop/HighLevelAPI.html)
-
-            if boil_off_position == None and self.cooling_jacket.coolant_transport.check_liquid(T = T_coolant[i], p = self.cooling_jacket.p0(x)) == False:
-                print(f"WARNING: Coolant boiled off at x = {x} m")
-                boil_off_position = x
-
-            #Get thermal circuit properties
-            q_dot[i], R_gas, R_wall, R_coolant, q_Adot[i] = self.thermal_circuit(x, h_gas[i], h_coolant[i], self.cooling_jacket.inner_wall, liner[i], T_gas[i], T_coolant[i])
-
-            #Calculate wall temperatures
-            T_wall_inner[i] = T_gas[i] - q_dot[i]*R_gas
-            T_wall_outer[i] = T_wall_inner[i] - q_dot[i]*R_wall
-
-        #Dictionary containing results
-        output_dict = {"x" : list(discretised_x),
-                "T_wall_inner" : list(T_wall_inner),
-                "T_wall_outer" : list(T_wall_outer),
-                "T_coolant" : list(T_coolant),
-                "T_gas" : list(T_gas),
-                "q_dot" : list(q_dot),
-                "q_Adot": list(q_Adot),
-                "h_gas" : list(h_gas),
-                "h_coolant" : list(h_coolant),
-                "boil_off_position" : boil_off_position}
-
-        #Export a .JSON file if required
-        if to_json != False:
-            with open(to_json, "w+") as write_file:
-                json.dump(output_dict, write_file)
-                print("Exported JSON data to '{}'".format(to_json))
-
-        return output_dict
-
-    def run_stress_analysis(self, heating_result, type="thermal", condition="steady"):
         """Perform stress analysis on the liner, using a cooling result.
         Args:
             heating_result (dict): Requires a heating analysis result to compute stress.
